Add skeleton loading UI, centralize Supabase anon key, and misc iOS/Android updates
- Replace spinner with animated skeleton UI on summary and quiz tabs (iOS + Android)
- Move Supabase anon key out of hardcoded strings into Secrets.supabaseAnonKey (iOS)
  and BuildConfig.SUPABASE_ANON_KEY (Android) across StudyCache, FeedManager, ShiurClient
- Delete stale orphaned StudySessionManager.swift from repo root (was not compiled)
- Various in-progress iOS and Android updates across Resources, Bookmarks, Settings,
  YCTLibraryClient, RSSParser, and related ViewModels

+++ android/app/src/main/java/com/anydaf/data/db/BookmarkEntity.kt
@@ -12,7 +12,7 @@ data class BookmarkEntity(
     val name: String,
     val notes: String,
     val tractateIndex: Int,
-    val daf: Int,
+    val daf: Double,
     val amud: Int,
     val studySectionIndex: Int?,
     val createdAt: Long  // epoch millis
